add server roles

+++ D365FFo/DevT1/DevTools/RestoreAxDB.ps1
@@ -1,6 +1,6 @@
 ﻿[cmdletbinding()]
 Param (
- [Parameter(Mandatory=$True)]
+    [Parameter(Mandatory = $True)]
     [string[]] $BackupFilePath
 )
 
@@ -11,18 +11,18 @@ Import-Module d365fo.tools
 $NewDBName = "AxDB"
 $Date = Get-Date -format "yyyyMMdd"
 $timenow = Get-date -Format "HHmm"
-$name = $NewDBName + "_" + $Date+ "_" + $timenow
+$name = $NewDBName + "_" + $Date + "_" + $timenow
 
 
 $BackupInfo = Get-DbaBackupInformation -Path $BackupFilePath -SqlInstance .
-$BackupInfo.FileList | %{If ($_.Type -eq "D") { $LogicaData = $_.LogicalName}
-Elseif ($_.Type -eq "L") {$LogicalLog = $_.LogicalName}
+$BackupInfo.FileList | % { If ($_.Type -eq "D") { $LogicaData = $_.LogicalName }
+    Elseif ($_.Type -eq "L") { $LogicalLog = $_.LogicalName }
 }
 
 $FileStructure = @{
- "$LogicaData" = "C:\Program Files\Microsoft SQL Server\MSSQL13.MSSQLSERVER\MSSQL\DATA\$name.mdf"
- "$LogicalLog" = "C:\Program Files\Microsoft SQL Server\MSSQL13.MSSQLSERVER\MSSQL\DATA\$name.ldf"
- }
+    "$LogicaData" = "C:\Program Files\Microsoft SQL Server\MSSQL13.MSSQLSERVER\MSSQL\DATA\$name.mdf"
+    "$LogicalLog" = "C:\Program Files\Microsoft SQL Server\MSSQL13.MSSQLSERVER\MSSQL\DATA\$name.ldf"
+}
 
 #restore DB from File
 Restore-DbaDatabase -SqlInstance . -Path $BackupFilePath -DatabaseName $name -ReplaceDbNameInFile -FileMapping $FileStructure -WithReplace
@@ -34,7 +34,7 @@ Stop-D365Environment
 #check if the Orig DB is present and remove it
 
 If (Get-DbaDatabase -SqlInstance . -Database Axdb_original) {
-Remove-DbaDatabase -SqlInstance . -Database Axdb_original -Confirm:$false
+    Remove-DbaDatabase -SqlInstance . -Database Axdb_original -Confirm:$false
 }
 
 #Change active DB
@@ -45,24 +45,30 @@ Switch-D365ActiveDatabase -DatabaseServer . -DatabaseName AxDB -SourceDatabaseNa
 $users = Get-LocalUser
 
 #get the email addresses
-$emails = $users | where {$_.Description -like  "*@*"} | select -expand Description
+$emails = $users | where { $_.Description -like "*@*" } | select -expand Description
 
 #g
 # create LocalUsers and provision
 foreach ($email in $emails) {
 
-$pieces = $Email -split '@'
+    $pieces = $Email -split '@'
 
-$FullName = $pieces[0] -replace "\.", " " 
-$Names = $pieces[0] -split "\." 
+    $FullName = $pieces[0] -replace "\.", " " 
+    $Names = $pieces[0] -split "\." 
 
-$name = ($names[0])[0] + $Names[1] 
+    $name = ($names[0])[0] + $Names[1] 
 
-#remove users before reimport
-Remove-D365User  -Email $email -ErrorAction SilentlyContinue
+    #remove users before reimport if already present to maintent the worker. 
+    $D365user = Get-D365User -Email $email
+    If ($D365user) {
+        Remove-D365User -Email $email -ErrorAction SilentlyContinue
 
-#add user in AX instance
-Import-D365ExternalUser -Id $Name -Name $FullName -Email $Email -DatabaseServer . -DatabaseName AxDB
+        #add user in AX instance
+        Import-D365ExternalUser -Id $D365user.UserId -Name $FullName -Email $Email -DatabaseServer . -DatabaseName AxDB
+    }
+    else {
+        Import-D365ExternalUser -Id $pieces[0] -Name $FullName -Email $Email -DatabaseServer . -DatabaseName AxDB
+    }
 }
 
 #Start ENV
